backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
google-auth-oauthlib==1.1.0
google-auth==2.23.3
google-api-python-client==2.108.0
firebase-admin==6.2.0
python-dotenv==1.0.0
nltk==3.8.1
pydantic==2.5.2
requests==2.31.0 
pandas==2.2.1
numpy==1.26.4
transformers==4.51.3
torch==2.0.1
scikit-learn==1.4.1 